Added a system for mods (WIP)
This is what I have so far. I didn't want to go any farther then get a no on the idea. So this way there are a ton of random planets all throughout the system. It's kind of cluttered, but also makes it feel like you're in a shopping mall with a ton of cool stores around you. From here it would just be a matter of going to each mod and collecting all the outfits and ships from them.